requirements.txt
django==1.8.3
MySQL-python==1.2.5
pymongo==3.2.2
mkdocs>=0.14.0
djangorestframework==3.3.3
django-ckeditor==5.0.3
pillow==3.2.0
djangorestframework-jwt==1.7.2
django-filter==0.13.0
